# packages/api-spec/src/crate-spec.openapi.yml
openapi: 3.0.3
info:
  version: "1.0.0"
  title: "Crate Network API"
  description: |
    ## About this spec

    The Crate API exposes several endpoints which allows the Crate macOS, iOS, 
    iPadOS and web clients to interact with Crate's IPFS cluster without the 
    clients needing to run IPFS clients of their own. Crate Network implements 
    the IPFS Pinning API specification at the `/pins` endpoint, as described by 
    the spec found
    [on GitHub](https://github.com/ipfs/pinning-services-api-spec/blob/3196b0b466752a3626b716969703cbd48cb9bcf7/ipfs-pinning-service.yaml). 
    The pinning API accepts a Bearer authorization token retrieved from your
    Crate account. This identifies your user and adds the pinned content to 
    your files on Crate. This version of the API is modified to include 
    metadata, all of which is optional so that Crate continues to conform to 
    the IPFS spec.

    The Filecoin replication service is handled via the pinning API. Instead of 
    using untrusted centralized cloud storage as the backend, Crate safely 
    replicates content onto other decentralized networks. When a request comes 
    to the pinning service to retrieve the content, Crate will prioritize 
    cached content, otherwise falling back to the Filecoin network to retrieve
    your files.
servers:
  - url: https://crate.network/api/v1
paths:
  /file:
    get:
      summary: Fetches FileModel for a path relative to the user root.
      parameters:
        - name: path
          in: query
          required: false
          schema:
            description: File path
            example: /dir1/dir2/file.txt
            type: string
      description: |
        Get FileModel object a file/directory specified by either a CID or
        the path from the root of the user's filesystem. If neither are 
        specified, we return the FileModel for the root directory.
      tags:
        - files
      responses:
        "200":
          description: Successful response (FileModel object)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileModel"
    post:
      parameters:
        - name: path
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CID"
      description: |
        Upload one or many files/directories to a path specified in the query
        parameters. If ignored, this will save the files to the root directory.
      tags:
        - files
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Successful response (FileModel object)
          content:
            application/json:
              schema:
                type: array
                minItems: 1
                items:
                  $ref: "#/components/schemas/FileModel"
...
components:
  schemas:
    FileModel:
      description: A file object as represented on @crate/web
      type: object
      required:
        - cid
        - type
        - size
        - date
      properties:
        cid:
          $ref: "#/components/schemas/CID"
        name:
          description: File/folder name
          type: string
          example: dir
        type:
          description: Whether this is a file or directory
          type: string
          enum:
            - "directory"
            - "file"
        signedEncryptionKey:
          description: |
            Encryption key for the file in string format, parsed from binary
            representation in UTF-8. This key has been encrypted by the user's
            data key. If this exists, it means the file is for private
            consumption.
          type: string
        size:
          description: The size of the file.
          type: number
          example: 1048576
        date:
          description: The date that this file was last modified in ISO-8601.
          type: string
          example: 2022-06-23T22:45:19.790Z
        links:
          description: |
            The CIDs linked to from this directory.
          type: array
          items:
            description: |
              Structure for linked items in a directory. Name can vary from 
              directory to directory.
            type: object
            required:
              - "cid"
              - "name"
              - "size"
            properties:
              cid:
                $ref: "#/components/schemas/CID"
              name:
                description: Name of linked file.
                type: string
                maxLength: 255
                example: file.txt
              size:
                description: Size of linked file.
                type: number
                example: 145
        cumulativeSize:
          description: The size of the directory and all contained files.
          type: number
    CID:
      description: Content Identifier (CID)
      type: string
      example: "Qm..."
...
  securitySchemes:
    accessToken:
      description: |
        An opaque token is required to be sent with each request in the HTTP 
        header:
        - `Authorization: Bearer <access-token>`
        The `access-token` should be generated per device, and the user should 
        have the ability to revoke each token separately.
      type: http
      scheme: bearer
security:
  - accessToken: []